requirements.txt
# requirements.txt
discord.py==2.4.0
python-dotenv==0.20.0
requests==2.27.1
sqlalchemy==2.0.36
pydantic==2.9.2
beautifulsoup4==4.12.3
